Fixed some faults
Lowered delta_t to not skip the necessary moments. To avoid long time of graphics building, now take only every 100 point to the arrays that get plotted, added some necessary things to the calculate_hit() function to capture the t_end moment, round the coordinates to 6 digits after the dot. Turned off check_t_not_negative(), since I seem to have not gotten the correct equations, now it's just checking that t_1 and t_0 are both greater or equal than 0. Added a simple graphic interface with tkinter and checking that all the parameters are entered correctly for the function to work. Rewrote the report accordingly, Should be done now.